StyledText: move it from commands to core package
As the updated UML shows (https://cloud.githubusercontent.com/assets/1260727/15242017/ca119324-18f3-11e6-8a3c-1d4fdae7997e.png),
too many classes/packages depend on it. It is best placed into the core package
for this reason.
