+++ org.metaborg.spoofax.shell.console/src/test/java/org/metaborg/spoofax/shell/commands/StyledTextTest.java
@@ -9,6 +9,7 @@ import java.awt.Color;
 import org.junit.Test;
 import org.metaborg.core.style.RegionStyle;
 import org.metaborg.core.style.Style;
+import org.metaborg.spoofax.shell.core.StyledText;
 
 /**
  * Test cases for the {@link StyledText} class, which wraps Spoofax {@link RegionStyle}.
