+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/IDisplay.java
@@ -1,6 +1,6 @@
 package org.metaborg.spoofax.shell.client;
 
-import org.metaborg.spoofax.shell.commands.StyledText;
+import org.metaborg.spoofax.shell.core.StyledText;
 
 /**
  * An {@link IDisplay} is for displaying the results of a command or expression that was executed,
