+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/commands/SpoofaxEvaluationCommand.java
@@ -2,6 +2,8 @@ package org.metaborg.spoofax.shell.commands;
 
 import java.util.function.Consumer;
 
+import org.metaborg.spoofax.shell.core.StyledText;
+
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
