+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/core/StyledText.java
@@ -1,4 +1,4 @@
-package org.metaborg.spoofax.shell.commands;
+package org.metaborg.spoofax.shell.core;
 
 import java.awt.Color;
 import java.util.List;
